% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_parties}
\alias{can_parties}
\title{Canada Federal Parties}
\format{
A character vector of 6 elements.
}
\usage{
data(can_parties)
}
\description{
A vector of Canada federal party codes.
}
\keyword{datasets}
