% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{can_party_colors}
\alias{can_party_colors}
\title{Canada Parties' Colors}
\format{
A named vector of 5 elements.
}
\usage{
data(can_party_colors)
}
\description{
A named vector mapping party codes to their colors.
}
\keyword{datasets}
