% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_issues}
\alias{clean_issues}
\title{Clean Issues}
\format{
A named vector of 8 elements.
}
\usage{
data(clean_issues)
}
\description{
A named vector mapping short issue codes to their descriptions.
}
\keyword{datasets}
