% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_and_party_positions.R
\name{compute_party_scales}
\alias{compute_party_scales}
\title{Compute Party Scales from Aggregated Issue Positions}
\usage{
compute_party_scales(aggregated_issue_positions, factanal_object, scale_names)
}
\arguments{
\item{aggregated_issue_positions}{A data frame of aggregated issue positions for parties.
Expected to be the output from `potgrowth::aggregate_party_positions`.}

\item{factanal_object}{A factor analysis object containing loadings for scales.
Expected to have loadings corresponding to the scales specified in `scale_names`.}

\item{scale_names}{A character vector specifying the names of the scales in the order
they appear in the factor analysis object loadings.}
}
\value{
A matrix with parties as row names and scales as column names, containing
        scale scores for each party.
}
\description{
This function takes aggregated issue positions for different parties, a factor analysis object,
and a vector of scale names. It returns a matrix of scale scores for each party based on
the specified scales in the factor analysis object.
}
\examples{
# Assuming `aggregated_issue_positions` and `factanal_object` are defined:
scale_names <- c("souv_langfr", "libertarian", "woke", "laicite", "immigration", "lien3")
party_scales <- compute_party_scales(aggregated_issue_positions,
                                     factanal_object,
                                     scale_names)
print(party_scales)
}
