% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_and_party_positions.R
\name{compute_respondents_saliency}
\alias{compute_respondents_saliency}
\title{Compute Respondents' Saliency on Scales and Parties}
\usage{
compute_respondents_saliency(
  data,
  rci_columns_prefix = "irc_",
  attitudegap_columns_prefix = "attitudegap_",
  scale_columns_prefix = "scale",
  remove_scale_leftRight = TRUE,
  parties = potgrowth::qc_parties
)
}
\arguments{
\item{data}{A dataframe containing the dataset to be analyzed.}

\item{rci_columns_prefix}{A string specifying the prefix of the RCI columns in `data`.}

\item{attitudegap_columns_prefix}{A string specifying the prefix of the attitude gap columns in `data`.}

\item{scale_columns_prefix}{A string specifying the prefix for scale columns in `data` from which the attitude gaps are calculated.}

\item{remove_scale_leftRight}{A logical indicating whether to exclude the left-right scale from calculations.}

\item{parties}{A vector of strings specifying the parties for which the analysis is to be performed.}
}
\value{
A matrix with the correlation coefficients between RCIs and attitude gaps for each respondent, indicating the saliency of issues.
}
\description{
This function computes the correlation between Relative Confidence Indices (RCIs) and attitude gaps for respondents, indicating the saliency of different issues among different parties.
}
\examples{
saliency <- compute_respondents_saliency(my_data, "irc_", "attitudegap_", "scale", TRUE, c("party1", "party2"))
}
