% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models.R
\name{dynamic_potgrowth_data}
\alias{dynamic_potgrowth_data}
\title{Dynamically Model and Predict Potential Growth Data with SES Interactions}
\usage{
dynamic_potgrowth_data(
  data,
  parties,
  issues,
  ses_interactions = character(0),
  model_variables = c("region", "age_cat", "income", "male", "educ", "lang", "religion")
)
}
\arguments{
\item{data}{A dataframe containing the original dataset to be processed.}

\item{parties}{A vector of political party identifiers to be analyzed.}

\item{issues}{A vector of issue identifiers to be analyzed.}

\item{ses_interactions}{A character vector of up to two SES variable names to be used for interaction
with the 'position' variable in the models. Defaults to an empty character vector.}

\item{model_variables}{A character vector of additional model variables to include
in the regression models. Defaults to c("region", "age_cat", "income",
"male", "educ", "lang", "religion").}
}
\value{
A dataframe combining all the predictions across the specified parties
        and issues. Each row contains the estimates and confidence intervals
        for IRC and voter likelihood, along with party and issue information.
}
\description{
This function processes a dataset to model and predict potential growth
data for various political parties and issues, including optional SES interactions.
It uses both linear and logistic regression to predict IRC and voter likelihoods respectively.
The function processes each party and issue combination, generating a set
of predictions with confidence intervals. Supports interaction with up to two
SES variables.
}
\examples{
test <- dynamic_potgrowth_data(
  data = data,
  parties = c("CAQ", "QS"),
  issues = issues,
  ses_interactions = c("region"),
)

}
