% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{geom_irc}
\alias{geom_irc}
\title{Create Custom Geom for IRC Visualization}
\usage{
geom_irc(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()` or `aes_()`.
Expected aesthetics are `conf_low_irc`, `conf_high_irc`, `estimate_irc`,
`estimate_vote`, `conf_low_vote`, `conf_high_vote`, and `color`.}

\item{data}{Optional dataset to use for the layer.}

\item{...}{Additional parameters for fine-tuning the `geom_linerange` components,
such as `size` and `alpha`.}
}
\value{
A ggplot2 layer.
}
\description{
This function creates a custom geom designed specifically for visualizing
the IRC along with its confidence intervals and other related estimates.
It adds multiple `geom_linerange` layers to a ggplot object, using color as an aesthetic.
}
\examples{
df <- data.frame(
  conf_low_irc = rnorm(10, 0.1, 0.05),
  conf_high_irc = rnorm(10, 0.2, 0.05),
  estimate_irc = rnorm(10, 0.15, 0.05),
  estimate_vote = runif(10, 0.4, 0.6),
  conf_low_vote = runif(10, 0.35, 0.45),
  conf_high_vote = runif(10, 0.55, 0.65),
  color = rep(c("red", "blue"), each = 5)
)
ggplot(df, aes(conf_low_irc = conf_low_irc, conf_high_irc = conf_high_irc,
               estimate_irc = estimate_irc, estimate_vote = estimate_vote,
               conf_low_vote = conf_low_vote, conf_high_vote = conf_high_vote,
               color = color)) +
  geom_irc()
}
