% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{get_quarto_graph}
\alias{get_quarto_graph}
\title{Generate a Quarto Graph}
\usage{
get_quarto_graph(
  survey_data,
  issue_slug,
  party_positions,
  choices,
  xlabel,
  remove_variables_from_models = NULL
)
}
\arguments{
\item{survey_data}{A data frame containing the survey data. It should include columns for `issue` and `position`.}

\item{issue_slug}{A string indicating the issue identifier.}

\item{choices}{A named vector of choices for the positions.}

\item{xlabel}{A string for the x-axis label.}
}
\value{
A Plotly graph object.
}
\description{
This function generates a Plotly graph for a given issue and survey data, with specific handling for the "iss_nationalisme_souv" issue.
}
\examples{
survey_data <- data.frame(issue = c("iss_nationalisme_souv", "iss_other"),
                          position = c(0.25, 0.75),
                          other_columns = c(1, 2))
choices <- c("0.25" = "Low", "0.75" = "High")
xlabel <- "Position"
get_quarto_graph(survey_data, "iss_nationalisme_souv", choices, xlabel)
}
