% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{get_quarto_party_dashboard}
\alias{get_quarto_party_dashboard}
\title{Generate Quarto Party Dashboard}
\usage{
get_quarto_party_dashboard(
  party,
  survey_data,
  issues_df,
  party_positions,
  remove_variables_from_models,
  width = 700
)
}
\arguments{
\item{party}{A character vector indicating the party.}

\item{survey_data}{A data frame containing the survey data.}

\item{issues_df}{A data frame containing issue slugs, labels, and position labels.}

\item{party_positions}{A data frame containing party positions on issues.}

\item{remove_variables_from_models}{A character vector of variables to be removed from the model.}
}
\value{
A `plotly` object representing the party dashboard.
}
\description{
This function generates a party dashboard using `plotly` based on survey data, issues data frame, and party positions.
}
\examples{
\dontrun{
party <- "PartyName"
survey_data <- data.frame(...) # Provide your survey data here
issues_df <- data.frame(issue_slug = c("issue1", "issue2"), issue_label = c("Issue 1", "Issue 2"), label0 = c("Less", "Less"), label1 = c("More", "More"))
party_positions <- data.frame(...) # Provide your party positions data here
remove_variables_from_models <- c("variable1", "variable2")
p <- get_quarto_party_dashboard(party, survey_data, issues_df, party_positions, remove_variables_from_models)
p
}
}
