% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_models.R
\name{get_voteint_model_data}
\alias{get_voteint_model_data}
\title{Get Data for Vote Intention Model}
\usage{
get_voteint_model_data(
  data,
  voteint_variable = "voteInt",
  attitudegap_prefix = "attitudegap",
  saliency_prefix = "adj_saliency",
  ses_and_vis_to_include
)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{voteint_variable}{The name of the vote intention variable in the dataset.}

\item{attitudegap_prefix}{A character string specifying the prefix for attitude gap columns.}

\item{saliency_prefix}{A character string specifying the prefix for saliency columns.}

\item{ses_and_vis_to_include}{Variables to include in the final dataset, passed as unquoted names. These should
be socio-economic status (SES) and visibility (vis) variables that the user wishes to retain for analysis.}
}
\value{
A dataframe containing the selected columns based on the specified criteria.
}
\description{
This function filters and selects columns from a dataset based on criteria relevant for
analyzing vote intention. It selects the vote intention variable, attitude gap variables,
saliency variables, and any additional socio-economic status (SES) and visibility (vis)
variables specified by the user.
}
\examples{
# Assuming 'your_data' is your dataset and it contains variables for vote intention,
# attitude gaps, saliency, and specific SES and visibility variables you are interested in:
get_voteint_model_data(data = your_data,
                       voteint_variable = "voteInt",
                       ses_and_vis_to_include = c("age", "education"))
}
