% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloc_recursive_frequentist.R
\name{inverse_gaussian_model}
\alias{inverse_gaussian_model}
\title{Fits an Inverse Gaussian regression model to data and plots the estimated and real density}
\usage{
inverse_gaussian_model(formula, link = "log", data, plot = TRUE)
}
\arguments{
\item{formula}{A formula specifying the response and predictor variables.}

\item{link}{The link function to use for the model. By default, the "identity" link is used.}

\item{data}{A dataframe containing the variables in the formula.}

\item{plot}{A logical. If TRUE, the function plots the estimated and real density. If FALSE, no plot is produced.}
}
\description{
Fits an Inverse Gaussian regression model to data and plots the estimated and real density
}
