% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_models.R
\name{iv_model}
\alias{iv_model}
\title{Independant Variable (IV) Linear Model}
\usage{
iv_model(data, iv_to_predict, ses)
}
\arguments{
\item{data}{A dataframe containing the dataset for the analysis.}

\item{iv_to_predict}{A string specifying the name of the independant variable that
the model will predict. This variable should be present in `data`.}

\item{ses}{A character vector of names of socio-economic status (SES) variables
included as predictors in the model. These variables should be present in `data`.}
}
\value{
An object of class `lm` representing the fitted linear model.
}
\description{
Fits a linear model to predict a specified independant variable (IV) based on
one or more socio-economic status (SES) variables provided by the user. This function
is designed to facilitate the creation of linear models by abstracting away the formula
creation process and directly using column names provided as strings.
}
\examples{
# Assuming 'df' is your dataframe, 'attitude_on_immigration' is the independant variable, and
# 'age' and 'education' are SES variables:
model <- iv_model(data = df, iv_to_predict = "attitude_on_immigration", ses = c("age", "education"))
summary(model)
}
