% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_models.R
\name{lm_with_interactions}
\alias{lm_with_interactions}
\title{Linear Model with Interaction Terms Between Atittude Gaps and Issue Saliency}
\usage{
lm_with_interactions(
  data,
  vd = "rci",
  attitudegap_prefix = "attitudegap",
  saliency_prefix = "adj_saliency"
)
}
\arguments{
\item{data}{A dataframe containing the dataset for analysis.}

\item{vd}{The name of the dependent variable in the model.}

\item{attitudegap_prefix}{A character string specifying the prefix for attitude gap columns.}

\item{saliency_prefix}{A character string specifying the prefix for saliency columns.}
}
\value{
A linear model object containing the fitted model with interaction terms.
}
\description{
Constructs and fits a linear model using a specified dependent variable and
interaction terms between attitude gap variables and saliency variables.
It automatically identifies variables starting with specified prefixes for
attitude gaps and saliency, creates interaction terms between each pair of
corresponding attitude and saliency variables, and includes these interactions
in the linear model along with other variables included in data.
}
\examples{
# Assuming 'your_data' is a dataframe that contains a dependent variable 'rci',
# and multiple attitude gap and saliency variables with respective prefixes:
model <- lm_with_interactions(data = your_data,
                              vd = "rci",
                              attitudegap_prefix = "attitudegap",
                              saliency_prefix = "adj_saliency")
summary(model)
}
