% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloc_recursive_frequentist.R
\name{lm_with_residuals}
\alias{lm_with_residuals}
\title{Generate diagnostic plots for a linear model}
\usage{
lm_with_residuals(formula, data)
}
\arguments{
\item{formula}{A formula specifying the response and predictor variables.}

\item{data}{A dataframe containing the variables in the formula.}
}
\value{
A linear model object with a diagnostic_plot attribute.
}
\description{
This function creates a linear model from a given formula and data, generates
diagnostic plots using autoplot, and adds the plot to the model object as an
attribute.
}
\examples{
# Example with mtcars dataset
data(mtcars)
model <- lm_with_residuals(mpg ~ wt, data = mtcars)

}
