% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models.R
\name{mode}
\alias{mode}
\title{Find the Mode of a Factor Variable}
\usage{
mode(x)
}
\arguments{
\item{x}{A factor variable. The function will stop and return an error if 'x' is not a factor.}
}
\value{
A vector containing the mode(s) of the factor. If there are multiple modes,
all of them are returned.
}
\description{
This function calculates the mode of a factor variable, i.e., the level
that appears most frequently in the factor. It returns the mode or modes if
there are ties for the most frequent level.
}
\examples{
factor_variable <- factor(c("red", "blue", "red", "green", "blue", "red"))
mode(factor_variable)
}
