% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_models.R
\name{multinom_with_interactions}
\alias{multinom_with_interactions}
\title{Multinomial Logistic Regression with Interaction Terms}
\usage{
multinom_with_interactions(
  data,
  parties,
  vd = "voteInt",
  attitudegap_prefix = "attitudegap",
  saliency_prefix = "adj_saliency"
)
}
\arguments{
\item{data}{A dataframe containing the dataset for analysis.}

\item{parties}{A character vector of party names to be used in creating interaction terms.}

\item{vd}{The name of the vote intention variable in the dataset.}

\item{attitudegap_prefix}{A character string specifying the prefix for attitude gap columns.}

\item{saliency_prefix}{A character string specifying the prefix for saliency columns.}
}
\value{
A multinomial logistic regression model object from the `nnet` package.
}
\description{
This function prepares and fits a multinomial logistic regression model using the `nnet` package,
with a specified vote intention variable as the dependent variable. It automatically creates interaction terms
between attitude gap variables and saliency variables for specified parties, incorporating these interactions
into the model.
}
\examples{
# Assuming 'your_data' is a dataframe that contains a vote intention variable,
# attitude gap variables, and saliency variables for different parties:
parties <- c("Party1", "Party2")
model <- multinom_with_interactions(data = your_data,
                                    parties = parties,
                                    vd = "voteInt",
                                    attitudegap_prefix = "attitudegap",
                                    saliency_prefix = "adj_saliency")
summary(model)

}
