% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{parse_money_vector}
\alias{parse_money_vector}
\title{Parse Money Vector}
\usage{
parse_money_vector(values, sep, floor, ceiling, ceiling_increment = 10000)
}
\arguments{
\item{values}{A vector of monetary values as strings to be parsed.}

\item{sep}{A character string specifying the separator used in monetary ranges.}

\item{floor}{The floor value to be specifically handled.}

\item{ceiling}{The ceiling value to be specifically handled.}

\item{ceiling_increment}{A numeric value to increment ceilings by, defaults to 10000.}
}
\value{
A numeric vector where each element is the parsed and rounded mean of the corresponding input value.
}
\description{
This function processes a vector of monetary values, parsing each value into a numeric form. The function handles
specific floor and ceiling values differently by using specified limits and an increment for ceilings.
}
\examples{
# Example usage:
# parse_money_vector(values = c("1000-2000", "floor", "ceiling"), sep = "-", floor = "floor", ceiling = "ceiling")

}
