% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_parties}
\alias{qc_parties}
\title{QC Provincial Parties}
\format{
A character vector of 5 elements.
}
\usage{
data(qc_parties)
}
\description{
A vector of QC provincial party codes.
}
\keyword{datasets}
