% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_and_party_positions.R
\name{read_one_file}
\alias{read_one_file}
\title{Read a Single ODS File and Append Filename as a Column}
\usage{
read_one_file(path)
}
\arguments{
\item{path}{The file path to the ODS file to be read.}
}
\value{
A data frame with the contents of the ODS file and an additional
        column named `coder` containing the filename from which the data was read.
}
\description{
This function reads a single ODS file from the specified path,
appends the filename (without extension) as a new column to the data,
and returns the modified data frame. The filename is used to identify
the source of the data within the returned data frame.
}
\examples{
# Assuming you have an ODS file named "data_file.ods" in your working directory:
path <- system.file("extdata", "data_file.ods", package = "YourPackageName")
data <- read_one_file(path)

}
