% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_and_party_positions.R
\name{retrieve_attitudegaps}
\alias{retrieve_attitudegaps}
\title{Retrieve Attitude Gaps for Parties}
\usage{
retrieve_attitudegaps(
  data,
  attitudegap_columns_prefix = "attitudegap_",
  scale_columns_prefix = "scale",
  parties = potgrowth::qc_parties,
  remove_scale_leftRight = TRUE
)
}
\arguments{
\item{data}{A dataframe containing the dataset from which attitude gaps are to be calculated.}

\item{attitudegap_columns_prefix}{A string specifying the prefix of the columns in `data` related to attitude gaps.}

\item{scale_columns_prefix}{A string specifying the prefix for scale columns in `data` from which the attitude gaps are calculated. Defaults to "scale".}

\item{parties}{A vector of strings specifying the parties for which attitude gaps are to be calculated.}

\item{remove_scale_leftRight}{A logical indicating whether to exclude the left-right scale from calculations. Defaults to TRUE.}
}
\value{
A list of matrices, each representing the attitude gaps for a row in `data`, structured by the specified scales and parties.
}
\description{
This function calculates the attitude gaps for specified parties within given scales, excluding optionally the left-right scale. An attitude gap measures the difference in attitudes across different dimensions.
}
\examples{
attitude_gaps <- retrieve_attitudegaps(my_data, "attitudegap_", "scale", c("party1", "party2"), TRUE)
}
