% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{geom_irc}
\alias{geom_irc}
\title{Create a Composite Geom for IRC and Vote Estimates}
\usage{
geom_irc(
  conf_low_irc,
  conf_high_irc,
  estimate_irc,
  conf_low_vote,
  conf_high_vote,
  estimate_vote,
  irc_conf_int_color = "black",
  ...
)
}
\arguments{
\item{conf_low_irc}{Lower bound of the IRC confidence interval.}

\item{conf_high_irc}{Upper bound of the IRC confidence interval.}

\item{estimate_irc}{Point estimate for IRC.}

\item{conf_low_vote}{Lower bound of the vote confidence interval.}

\item{conf_high_vote}{Upper bound of the vote confidence interval.}

\item{estimate_vote}{Point estimate for the vote.}

\item{irc_conf_int_color}{Color for the IRC confidence interval. Default is "black".}

\item{...}{Additional parameters passed to `geom_linerange`.}
}
\value{
A list of `geom_linerange` objects for use with ggplot2.
}
\description{
This function creates a list of `geom_linerange` elements to represent the
confidence intervals and estimates for IRC and vote values.
}
\examples{
ggplot(data, aes(x = your_x_variable, y = your_y_variable)) +
  geom_irc(conf_low_irc = your_data$conf_low_irc,
           conf_high_irc = your_data$conf_high_irc,
           estimate_irc = your_data$estimate_irc,
           conf_low_vote = your_data$conf_low_vote,
           conf_high_vote = your_data$conf_high_vote,
           estimate_vote = your_data$estimate_vote,
           aes(color = your_color_variable))
}
