% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_models.R
\name{iv_model}
\alias{iv_model}
\title{Independent Variable (IV) Regression Model}
\usage{
iv_model(data, iv_to_predict, ses, model_type = "lm")
}
\arguments{
\item{data}{A dataframe containing the dataset for the analysis.}

\item{iv_to_predict}{A string specifying the name of the independent variable that
the model will predict. This variable should be present in `data`.}

\item{ses}{A character vector of names of socio-economic status (SES) variables
included as predictors in the model. These variables should be present in `data`.}

\item{model_type}{A string that specifies the type of regression model to fit.
Accepts "lm" for linear regression and "multinom" for multinomial logistic regression.
Defaults to "lm".}
}
\value{
An object representing the fitted model. The class of the object will
  depend on the `model_type` parameter: an object of class `lm` for linear models
  or class `nnet` for multinomial logistic regression models.
}
\description{
Fits a model to predict a specified independent variable (IV) based on
one or more socio-economic status (SES) variables provided by the user. This function
is designed to facilitate the creation of models by abstracting away the formula
creation process and directly using column names provided as strings. It supports
fitting both linear models (using ordinary least squares) and multinomial logistic
regression models.
}
\examples{
# Assuming 'df' is your dataframe, 'attitude_on_immigration' is the independent variable, and
# 'age' and 'education' are SES variables:
# For a linear model:
model_lm <- iv_model(data = df, iv_to_predict = "attitude_on_immigration", ses = c("age", "education"), model_type = "lm")
summary(model_lm)
# For a multinomial logistic regression model:
model_multinom <- iv_model(data = df, iv_to_predict = "attitude_on_immigration", ses = c("age", "education"), model_type = "multinom")
summary(model_multinom)
}
